logging: Remove global _stream object, add a root logger instead.
To allow formatting log messages a LogRecord will be created for each
logging activity. A LogRecord contains certain fields that can be used
by the Formatter to define the layout of the log messages.

In order to allow formatting the central _stream variable can not be
used anymore. Instead there is a "root" logger according to the CPython
logging module.